Add files via upload
It is a safety cap that is part of the anti-drowning system for long-distance drivers. The safety cap measures the angle of a person's head and calculates the average. If an average value is not equal to that of the average, it warns of sound and vibration. I wake you up by giving you a warning.